Añadiendo cambios al preview de pdf

+++ fish/.config/fish/functions/pdf_preview.fish
@@ -4,7 +4,10 @@ function pdf_preview --description="previsualizar pdfs en terminal"
 		set file "$( basename $(echo "Descargas/0.1_Presentación.pdf" | tr ' ' '_'))"
 		pdftoppm -f 1 -l 1 "$filename" >> "/tmp/$file.png"
 	end
-	 kitty +kitten icat --silent --stdin no --transfer-mode file "/tmp/$file.png" < /dev/null > /dev/tty
-	 rm -r "/tmp/$file.png"
+	if test -e "/tmp/$file.png"
+	 	kitty +kitten icat --silent --stdin no --transfer-mode file "/tmp/$file.png" < /dev/null > /dev/tty
+	 	rm -r "/tmp/$file.png"
+	end
+
 
  end
